+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.3.0] - 19 April 2020.
+### Breaking changes
+* Renamed `ApiBase` to `BaseApi`
+* Split `ApiRequest` to `ApiRequest` and abstract `BaseApiRequest` classes
+* Replaced `BaseApi` `call` method with `send` method which accepts the argument `BaseApiRequest`.
 ## [0.2.0] - 18 April 2020.
 * Add example
 * Fixed request duration bug (DebugLink)